% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialroulette.R
\name{frame_to_partition}
\alias{frame_to_partition}
\title{Take a frame and convert this to a corresponding partition}
\usage{
frame_to_partition(frame)
}
\arguments{
\item{frame}{The frame with `id` and `group` columns to convert}
}
\value{
A partition, i.e. a list of vectors, where each vector contains all individuals in the corresponding group
}
\description{
This is a small useful converter function for taking a tibble including a `group` and `id` column and convert it to a partition.
}
\examples{
p <- tibble::tibble(id=sprintf("id\%.02d",1:5), group=c(1,1,1,2,2))
socialroulette::frame_to_partition(p)
}
\keyword{internal}
