% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialroulette.R
\name{frame_to_partitionList}
\alias{frame_to_partitionList}
\title{Take a frame including a `group` column and convert this to a corresponding partitionList}
\usage{
frame_to_partitionList(frame)
}
\arguments{
\item{frame}{The frame with `id` and `group` columns to convert}
}
\value{
A list of vectors, where each vector contains all individuals in the corresponding group
}
\description{
This is a small useful converter function
}
