% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialroulette.R
\name{group_to_pairs}
\alias{group_to_pairs}
\title{Make all pairs within a group}
\usage{
group_to_pairs(group)
}
\arguments{
\item{group}{A vector of ids belonging to the same group}
}
\value{
a data.frame containing all pairs with id of the first being smaller than the id of the second entry
}
\description{
Make all pairs within a group
}
\keyword{internal‹}
