% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialroulette.R
\name{make_partition_srs}
\alias{make_partition_srs}
\title{Simple partitioning individuals into groups using simple random sampling}
\usage{
make_partition_srs(frame, m)
}
\arguments{
\item{frame}{tibble containing a column `id` of the sampling units (e.g. people)}

\item{m}{requested group size (integer). If m not a divisor of nrow(people) then put leftovers in existing groups}
}
\value{
A list of vectors where each vector containing the ids (row number in frame) of those belonging to the same group
}
\description{
Simple partitioning individuals into groups using simple random sampling
}
\keyword{internal}
