% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialroulette.R
\name{mdgp_partition_to_frame}
\alias{mdgp_partition_to_frame}
\title{Match a partition based on index (i.e. MGDP output) to ids in the frame}
\usage{
mdgp_partition_to_frame(mdgp_partition, frame)
}
\arguments{
\item{mdgp_partition}{A tibble containing idx and group columns (MDGP output)}

\item{frame}{A frame (i.e. a tibble with at least a column denoted `id`) to match the idx to. This has to be the frame used to generate the partition in the first place.}
}
\value{
The `frame` tibble augmented with an additional `group` column
}
\description{
Match a partition based on index (i.e. MGDP output) to ids in the frame
}
\keyword{internal}
