% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialroulette.R
\name{pairs_to_partition}
\alias{pairs_to_partition}
\title{Function to convert a data.frame of pairs to a partition}
\usage{
pairs_to_partition(pairs_df)
}
\arguments{
\item{pairs_df}{data.frame containing the pairs, i.e. it has columns date, id1 and id2}
}
\value{
A partitionList
}
\description{
Function to convert a data.frame of pairs to a partition
}
\keyword{internal}
