% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialroulette.R
\name{pairs_to_partitions}
\alias{pairs_to_partitions}
\title{Convert a pairs data.frame into a partition list}
\usage{
pairs_to_partitions(pairs)
}
\arguments{
\item{pairs}{A \code{data.frame} with columns \code{date} and \code{id1} and \code{id2}.}
}
\description{
The function works by splitting the pairs data.frame by date and then apply the internal
function pairs_to_partition to the resulting sub data.frame.
}
\examples{
partitions <- list("2021-04-21"=list(c("id02", "id03", "id04"), c("id01", "id05")),
                  "2021-04-28"=list(c("id03", "id04", "id05"), c("id01", "id02")))
partitions2 <- partitions \%>\% socialroulette:::partitions_to_pairs() \%>\%
                          socialroulette:::pairs_to_partitions()
all.equal(partitions, partitions2)
}
\seealso{
pairs_to_partition
}
