% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialroulette.R
\name{pairs_to_session}
\alias{pairs_to_session}
\title{Convert a pairs data.frame into a session list}
\usage{
pairs_to_session(pairs)
}
\arguments{
\item{pairs}{A \code{data.frame} with columns date and \code{id1} and \code{id2}.}
}
\description{
Convert a pairs data.frame into a session list
}
\examples{
sessions <- list("2021-04-21"=list(c("id02", "id03", "id04"), c("id01", "id05")),
                  "2021-04-28"=list(c("id03", "id04", "id05"), c("id01", "id02")))
sessions2 <- sessions \%>\% socialroulette:::sessions_to_pairs() \%>\%
                          socialroulette:::pairs_to_session()
all.equal(sessions, sessions2)
}
