% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialroulette.R
\name{partitions_to_pairs}
\alias{partitions_to_pairs}
\title{Convert a list of partitions into a data.frame with all pairs}
\usage{
partitions_to_pairs(partitions)
}
\arguments{
\item{partitions}{A list of past partitions, i.e. a named list where each entry is a partition and the names reflect the dates the partition were used}
}
\description{
The name of the list contain the date at which each partition was used.
}
\examples{
partitions <- list("2021-04-21"=list(c("id02", "id03", "id04"), c("id05", "id01")),
                  "2021-04-28"=list(c("id05", "id03", "id04"), c("id02", "id01")))
socialroulette::partitions_to_pairs(partitions)
}
