% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialroulette.R
\name{read_mdgp_solutionfile}
\alias{read_mdgp_solutionfile}
\title{Read output from the Lai and Hao (2016) MDGP solver}
\usage{
read_mdgp_solutionfile(file_name)
}
\arguments{
\item{file_name}{Name of the solution file}
}
\value{
data.frame with two clumns, idx (index) and group
}
\description{
Read output from the Lai and Hao (2016) MDGP solver
}
