% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialroulette.R
\name{sessions_to_distance}
\alias{sessions_to_distance}
\title{Convert a list of sessions into a MDGP distance metric}
\usage{
sessions_to_distance(current_frame, past_sessions)
}
\arguments{
\item{current_frame}{The current frame}

\item{past_sessions}{List of previous sessions}
}
\description{
A session is a data.frame containing the columns id1, id2 and date
}
\examples{
frame  <- tibble::tibble(id=sprintf("id\%.2d", 1:5), date=as.Date("2021-04-28"))
sessions <- list("2021-04-14"=list(c("id02", "id03", "id04"), c("id01", "id05")),
                  "2021-04-21"=list(c("id03", "id04", "id05"), c("id01", "id02")))
socialroulette:::sessions_to_distance(frame, sessions)
}
