% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialroulette.R
\name{sessions_to_pairs}
\alias{sessions_to_pairs}
\title{Convert a list of partitionLists into a data.frame with all pairs}
\usage{
sessions_to_pairs(sessions)
}
\arguments{
\item{sessions}{A list of past partitions, i.e. a list where each entry is a partitionList}
}
\description{
The name of the list contain the date at which each partition was used.
}
\examples{
sessions <- list("2021-04-21"=list(c("id02", "id03", "id04"), c("id05", "id01")),
                  "2021-04-28"=list(c("id05", "id03", "id04"), c("id02", "id01")))
socialroulette:::sessions_to_pairs(sessions)
}
