% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialroulette.R
\name{write_mdgp_specfile}
\alias{write_mdgp_specfile}
\title{Convert partition specification to Lai and Hao (2016) MDGP solver input specification.}
\usage{
write_mdgp_specfile(current_frame, past_sessions, m)
}
\arguments{
\item{current_frame}{A pair-distance tibble. The frame needs to contain a column denoted `id`}

\item{past_sessions}{A list of sessions. Each sessions consists of a session specification in list format.}

\item{m}{The minimum group size, i.e. all groups have size at least m}
}
\value{
File name of the generated specification file
}
\description{
Generate a \code{tempfile()} containing the appropriate specification formed from \code{current_frame} and \code{past_sessions}.
}
\details{
The specification format has the following header line
\deqn{n\> m\> ds\> n_1\> n_1\> \ldots n_G \> n_G}
Here, each group size \eqn{n_g, g=1,\ldots,G} is repeated twice, because the solver has the opportunity to specify a lower as well as an
upper boundary for each group size. The keyword `ds` for the solver in the above means that the groups can be of different sizes as determined by the respective \eqn{n_g}.
}
\examples{
frame  <- tibble::tibble(id=sprintf("id\%.2d", 1:5), date=as.Date("2021-04-28"))
past_sessions <- list("2021-04-21"=list(c("id02", "id03", "id04"), c("id01", "id05")))
spec_file <- socialroulette:::write_mdgp_specfile(frame, past_sessions=past_sessions, m=2)
cat(stringr::str_c(readLines(spec_file), collapse="\n"))
}
